requirements.txt
flask==2.3.3
flask-cors==4.0.0
sqlalchemy==2.0.23
apscheduler==3.10.4
feedgen==1.0.0
google-genai==0.6.0
pytz==2024.1